refactor: existing code
