+++ packages/core/src/App/Components/Layout/Header/dtrader-v2/account-actions-dtrader-v2.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
+import 'Sass/app/_common/components/account-switcher-dtrader-v2.scss';
 import { LabelPairedBellLgRegularIcon } from '@deriv/quill-icons';
 import { formatMoney, moduleLoader } from '@deriv/shared';
 import { Badge } from '@deriv-com/quill-ui';
-import 'Sass/app/_common/components/account-switcher-dtrader-v2.scss';
 import { LoginButton } from '../login-button';
 import { SignupButton } from '../signup-button';
 
@@ -26,8 +26,8 @@ export type TAccountInfoDTraderV2 = Omit<
     'toggleAccountsDialog' | 'is_acc_switcher_disabled' | 'is_acc_switcher_on'
 > & {
     is_dialog_on?: boolean;
-    toggleDialog: (value?: boolean | undefined) => void;
     is_disabled?: boolean;
+    toggleDialog: (value?: boolean | undefined) => void;
 };
 const AccountInfoDTraderV2 = React.lazy(
     () =>
@@ -41,65 +41,61 @@ const AccountInfoDTraderV2 = React.lazy(
         }>
 );
 
-const AccountActionsDTraderV2 = React.memo(
-    ({
-        acc_switcher_disabled_message,
-        account_type,
-        balance,
-        currency,
-        is_acc_switcher_on,
-        is_acc_switcher_disabled,
-        is_eu,
-        is_logged_in,
-        is_virtual,
-        notifications_count,
-        toggleAccountsDialog,
-        account_switcher_title,
-    }: TAccountActionsDTraderV2) => {
-        if (is_logged_in) {
-            return (
-                <React.Fragment>
-                    <AccountInfoDTraderV2
-                        acc_switcher_disabled_message={acc_switcher_disabled_message}
-                        account_type={account_type}
-                        balance={typeof balance === 'undefined' ? balance : formatMoney(currency, balance, true)}
-                        is_disabled={is_acc_switcher_disabled}
-                        is_eu={is_eu}
-                        is_virtual={is_virtual}
-                        currency={currency}
-                        is_dialog_on={is_acc_switcher_on}
-                        toggleDialog={toggleAccountsDialog}
-                        account_switcher_title={account_switcher_title}
-                    />
-                    <div className='notifications__wrapper'>
-                        {notifications_count ? (
-                            <Badge
-                                variant='notification'
-                                position='top-right'
-                                label={notifications_count.toString()}
-                                color='danger'
-                                size='sm'
-                                contentSize='sm'
-                            >
-                                <LabelPairedBellLgRegularIcon className='notifications__icon' />
-                            </Badge>
-                        ) : (
-                            <LabelPairedBellLgRegularIcon className='notifications__icon' />
-                        )}
-                    </div>
-                </React.Fragment>
-            );
-        }
-        // TODO: remove Login/out after
+const AccountActionsDTraderV2 = ({
+    acc_switcher_disabled_message,
+    account_switcher_title,
+    account_type,
+    balance,
+    currency,
+    is_acc_switcher_on,
+    is_acc_switcher_disabled,
+    is_eu,
+    is_logged_in,
+    is_virtual,
+    notifications_count,
+    toggleAccountsDialog,
+}: TAccountActionsDTraderV2) => {
+    if (is_logged_in) {
         return (
             <React.Fragment>
-                <LoginButton className='acc-info__button' />
-                <SignupButton className='acc-info__button' />
+                <AccountInfoDTraderV2
+                    acc_switcher_disabled_message={acc_switcher_disabled_message}
+                    account_switcher_title={account_switcher_title}
+                    account_type={account_type}
+                    balance={typeof balance === 'undefined' ? balance : formatMoney(currency, balance, true)}
+                    currency={currency}
+                    is_disabled={is_acc_switcher_disabled}
+                    is_eu={is_eu}
+                    is_virtual={is_virtual}
+                    is_dialog_on={is_acc_switcher_on}
+                    toggleDialog={toggleAccountsDialog}
+                />
+                <div className='notifications__wrapper'>
+                    {notifications_count ? (
+                        <Badge
+                            color='danger'
+                            contentSize='sm'
+                            label={notifications_count.toString()}
+                            position='top-right'
+                            size='sm'
+                            variant='notification'
+                        >
+                            <LabelPairedBellLgRegularIcon className='notifications__icon' />
+                        </Badge>
+                    ) : (
+                        <LabelPairedBellLgRegularIcon className='notifications__icon' />
+                    )}
+                </div>
             </React.Fragment>
         );
     }
-);
+    // TODO: remove Log in/out after dev testing
+    return (
+        <React.Fragment>
+            <LoginButton className='acc-info__button' />
+            <SignupButton className='acc-info__button' />
+        </React.Fragment>
+    );
+};
 
-AccountActionsDTraderV2.displayName = 'AccountActionsDTraderV2';
-
-export { AccountActionsDTraderV2 };
+export default AccountActionsDTraderV2;
