+++ packages/core/src/App/Components/Layout/Header/dtrader-v2/header-account-actions-dtrader-v2.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { observer, useStore } from '@deriv/stores';
-import { AccountActionsDTraderV2 } from './account-actions-dtrader-v2';
+import AccountActionsDTraderV2 from './account-actions-dtrader-v2';
 import { getAccountTitle } from './Utils/account-switcher-dtrader-v2-utils';
 
 const HeaderAccountActionsDTraderV2 = observer(() => {
