+++ packages/core/src/App/Containers/Layout/header/__tests__/show-notifications.spec.tsx
@@ -12,12 +12,6 @@ jest.mock('App/Components/Layout/Header', () => {
     };
 });
 
-jest.mock('@deriv-com/quill-ui', () => ({
-    Badge: jest.fn(() => {
-        return 'MockedBadge';
-    }),
-}));
-
 describe('ShowNotifications', () => {
     let store: TStores;
     beforeEach(() => {
