+++ packages/core/src/App/Containers/Layout/header/dtrader-header-wallets.tsx
@@ -9,7 +9,7 @@ import { AccountActionsWallets } from 'App/Components/Layout/Header/wallets/acco
 import platform_config from 'App/Constants/platform-config';
 import RealAccountSignup from 'App/Containers/RealAccountSignup';
 import SetAccountCurrencyModal from 'App/Containers/SetAccountCurrencyModal';
-import NewVersionNotification from 'App/Containers/new-version-notification';
+import NewVersionNotification from 'App/Containers/new-version-notification.jsx';
 import ToggleMenuDrawer from 'App/Components/Layout/Header/toggle-menu-drawer.jsx';
 import AccountsInfoLoaderWallets from 'App/Components/Layout/Header/wallets/accounts-info-loader-wallets';
 import TradersHubHomeButton from './traders-hub-home-button';
